start to restructure commands
